Renames projects to services to be able to support different projects later
