+++ fabfile.py
@@ -4,8 +4,8 @@ from invoke.config import Config
 from environments.surfpol import create_configuration_and_session
 from commands.postgres.fabric import setup_postgres_remote
 from commands.elastic.fabric import connect_elastic_cluster, push_indices_template
-from commands.projects.service.fabric import create_snapshot, restore_snapshot
-from commands.projects.harvester.fabric import connect_uwsgi, connect_flower, connect_with_shell
+from commands.services.service.fabric import create_snapshot, restore_snapshot
+from commands.services.harvester.fabric import connect_uwsgi, connect_flower, connect_with_shell
 
 
 service_environment, _ = create_configuration_and_session(
