+++ tasks.py
@@ -6,8 +6,8 @@ from commands.elastic.tasks import create_decompound_dictionary, push_decompound
 from commands.deploy import (prepare_builds, build, push, deploy, migrate, print_available_images,
                              print_running_containers)
 from commands.test import test_collection
-from commands.projects.service.invoke import import_snapshot, sync_upload_media, sync_category_filters
-from commands.projects.harvester.invoke import (import_dataset, harvest, clean_data, index_dataset_version,
+from commands.services.service.invoke import import_snapshot, sync_upload_media, sync_category_filters
+from commands.services.harvester.invoke import (import_dataset, harvest, clean_data, index_dataset_version,
                                                 dump_data, sync_harvest_content, generate_previews,
                                                 promote_dataset_version, extend_resource_cache)
